clad-backend/requirements.txt
pandas
numpy
scikit-learn
lightgbm
fastapi
uvicorn
joblib
httpx
python-dotenv
networkx
pydantic